Use u16 for strip y (#81)

+++ crates/sparse_primitives/src/render.rs
@@ -233,7 +233,7 @@ impl<KE: KernelExecutor> InnerContext<KE> {
                 continue;
             }
 
-            if strip.y() >= self.height as u32 {
+            if strip.y() >= self.height as u16 {
                 // Since strips are sorted by location, any subsequent strips will also be
                 // outside the viewport, so we can abort entirely.
                 break;
